requirements.txt
fastapi>=0.109.1,<1.0.0
uvicorn[standard]>=0.24.0,<1.0.0
openai>=1.3.0,<2.0.0
stripe>=7.4.0,<8.0.0
python-dotenv>=1.0.0,<2.0.0
pydantic[email]>=2.5.0,<3.0.0
httpx>=0.25.0,<1.0.0
sqlalchemy>=2.0.23,<3.0.0
alembic>=1.12.1,<2.0.0
prometheus-client>=0.19.0,<1.0.0
apscheduler>=3.10.4,<4.0.0

airbyte-api
cryptography
python-jose[cryptography]>=3.3.0,<4.0.0
passlib[bcrypt]>=1.7.4,<2.0.0
argon2-cffi>=25.0.0,<26.0.0
python-multipart>=0.0.6,<1.0.0
email-validator>=2.0.0,<3.0.0
